requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
black==25.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
cssselect2==0.8.0
Django==5.2.5
djangorestframework==3.16.1
html5lib==1.1
idna==3.10
iniconfig==2.1.0
lxml==6.0.0
mypy_extensions==1.1.0
oscrypto==1.3.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyMySQL==1.1.1
pypdf==5.9.0
python-bidi==0.6.6
python-dotenv==1.1.1
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vulture==2.14
webencodings==0.5.1
xhtml2pdf==0.2.17
django-environ>=0.11
drf-spectacular
drf-spectacular-sidecar
